refactor: rename fission-common to fission-core
