--- fission-common/README.md
@@ -1,1 +1,0 @@
-Volumes, Fission IPLD block storage.
