+++ fission-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fission-common"
+name = "fission-core"
 version = "0.1.0"
 description = ""
 keywords = []
@@ -9,7 +9,7 @@ license = "Apache-2.0"
 readme = "README.md"
 edition = "2021"
 rust-version = "1.69"
-documentation = "https://docs.rs/fission-common"
+documentation = "https://docs.rs/fission-core"
 repository = "https://github.com/fission-codes/fission-server"
 authors = ["Steven Vandevelde <icid.asset@gmail.com>"]
