+++ fission-core/src/authority/error.rs
